+++ com.onesignal.unity.ios/Runtime/OneSignalIOS.Interface.cs
@@ -56,6 +56,7 @@ namespace OneSignalSDK {
         
         [DllImport("__Internal")] private static extern string _getDeviceState();
         
+        [DllImport("__Internal")] private static extern void _setLogLevel(int logLevel, int alertLevel);
         [DllImport("__Internal")] private static extern void _setPrivacyConsent(bool consent);
         [DllImport("__Internal")] private static extern bool _getPrivacyConsent();
         [DllImport("__Internal")] private static extern void _setRequiresPrivacyConsent(bool required);